@@ -13,6 +13,20 @@
     <url>https://github.com/markoniemi/microservice-example.git</url>
   </scm>
   <url>https://github.com/markoniemi/microservice-example/wiki</url>
+  <properties>
+    <maven.compiler.source>1.8</maven.compiler.source>
+    <maven.compiler.target>1.8</maven.compiler.target>
+    <java.version>1.8</java.version>
+    <spring-cloud.version>2020.0.3</spring-cloud.version>
+    <spring.boot.version>2.4.8</spring.boot.version>
+    <jacoco.maven.plugin.version>0.8.7</jacoco.maven.plugin.version>
+    <bootstrap.version>4.6.0-1</bootstrap.version>
+    <jquery.version>3.6.0</jquery.version>
+    <maven.cargo.version>1.6.5</maven.cargo.version>
+    <selenium.version>3.13.0</selenium.version>
+    <docker.image.prefix>markoniemi</docker.image.prefix>
+    <timestamp>${maven.build.timestamp}</timestamp>
+  </properties>
   <modules>
     <module>discovery</module>
     <module>config-server</module>
@@ -20,25 +34,21 @@
     <module>user-repository</module>
     <module>user-application</module>
   </modules>
-  <properties>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-    <java.version>1.8</java.version>
-    <spring-cloud.version>Greenwich.SR3</spring-cloud.version>
-    <spring.boot.version>2.1.9.RELEASE</spring.boot.version>
-    <jacoco.maven.plugin.version>0.8.5</jacoco.maven.plugin.version>
-    <bootstrap.version>3.3.6</bootstrap.version>
-    <jquery.version>2.1.1</jquery.version>
-    <maven.cargo.version>1.6.5</maven.cargo.version>
-    <selenium.version>3.13.0</selenium.version>
-  </properties>
   <build>
+    <resources>
+      <resource>
+        <directory>src/main/resources</directory>
+        <filtering>true</filtering>
+      </resource>
+    </resources>
     <plugins>
       <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.8.0</version>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>11</source>
+          <target>11</target>
         </configuration>
       </plugin>
       <plugin>
