@@ -1,11 +1,12 @@
 package example;
 
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import java.util.Arrays;
 import java.util.Collections;
 
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
@@ -19,7 +20,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.ContextHierarchy;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import example.config.IntegrationTestConfig;
 import example.user.UserApplication;
@@ -29,7 +30,7 @@ import lombok.extern.log4j.Log4j2;
  * Base class for integration tests, enables running multiple tests
  * with @SpringBootTest
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = UserApplication.class, webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @ContextHierarchy(@ContextConfiguration(classes = IntegrationTestConfig.class))
 //@ActiveProfiles("local")
@@ -42,7 +43,7 @@ public class AbstractIntegrationTestBase {
 
     @Test
     public void dummy() {
-        Assert.assertTrue(true);
+        assertTrue(true);
     }
 
     protected String get(String url, MediaType mediaType) {
@@ -53,7 +54,7 @@ public class AbstractIntegrationTestBase {
         ResponseEntity<String> response = new TestRestTemplate().exchange(url, HttpMethod.GET,
                 new HttpEntity<>(headers), String.class);
         log.debug(response.getBody());
-        Assert.assertTrue(response.getStatusCode() == HttpStatus.OK);
+        assertTrue(response.getStatusCode() == HttpStatus.OK);
         return response.getBody();
     }
 
