@@ -1,13 +1,15 @@
 package example;
 
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assumptions.assumeFalse;
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
+
 import java.io.IOException;
 import java.util.List;
 
-import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.collections4.CollectionUtils;
 import org.apache.http.client.ClientProtocolException;
-import org.junit.Assert;
-import org.junit.Assume;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.cloud.client.ServiceInstance;
 
 import lombok.extern.log4j.Log4j2;
@@ -18,17 +20,17 @@ public class ConfigIT extends AbstractIntegrationTestBase {
 
     @Test
     public void localConfig() throws ClientProtocolException, IOException {
-        Assume.assumeFalse(isCloudConfigEnabled());
+        assumeFalse(isCloudConfigEnabled());
         String body = get(url + "/actuator/env", null);
-        Assert.assertTrue(body.contains("runtime.environment"));
-        Assert.assertTrue(body.contains("local"));
+        assertTrue(body.contains("runtime.environment"));
+        assertTrue(body.contains("local"));
     }
     @Test
     public void remoteConfig() throws ClientProtocolException, IOException {
-        Assume.assumeTrue(isCloudConfigEnabled());
+        assumeTrue(isCloudConfigEnabled());
         String body = get(url + "/actuator/env", null);
-        Assert.assertTrue(body.contains("runtime.environment"));
-        Assert.assertTrue(body.contains("remote"));
+        assertTrue(body.contains("runtime.environment"));
+        assertTrue(body.contains("remote"));
     }
 
     private String getUrl() {
