@@ -1,10 +1,13 @@
 package example;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
 import javax.annotation.Resource;
 
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import example.repository.user.Role;
 import example.repository.user.User;
@@ -14,7 +17,7 @@ public class UserRestControllerIT extends AbstractIntegrationTestBase {
     @Resource
     UserClient userClient;
 
-    @Before
+    @BeforeEach
     public void setUp() throws InterruptedException {
         waitForServiceRegistration();
     }
@@ -22,38 +25,38 @@ public class UserRestControllerIT extends AbstractIntegrationTestBase {
     @Test
     public void users() {
         User[] users = userClient.findAll();
-        Assert.assertNotNull(users);
-        Assert.assertEquals("admin", users[0].getUsername());
+        assertNotNull(users);
+        assertEquals("admin", users[0].getUsername());
     }
 
     @Test
     public void userByUsername() {
         User user = userClient.findByUsername("admin");
-        Assert.assertNotNull(user);
-        Assert.assertEquals("admin", user.getUsername());
+        assertNotNull(user);
+        assertEquals("admin", user.getUsername());
     }
 
     @Test
     public void findById() {
         User user = userClient.findById(1L);
-        Assert.assertNotNull(user);
-        Assert.assertEquals("admin", user.getUsername());
+        assertNotNull(user);
+        assertEquals("admin", user.getUsername());
     }
 
     @Test
     public void create() {
         User user = userClient.create(new User("user", "user", "email", Role.ROLE_USER));
-        Assert.assertNotNull(user);
-        Assert.assertEquals("user", user.getUsername());
+        assertNotNull(user);
+        assertEquals("user", user.getUsername());
         user = userClient.findById(user.getId());
-        Assert.assertNotNull(user);
+        assertNotNull(user);
         userClient.delete(user.getId());
         user = userClient.findById(user.getId());
-        Assert.assertNull(user);
+        assertNull(user);
     }
 
     @Test
     public void count() {
-        Assert.assertEquals(1L, userClient.count());
+        assertEquals(1L, userClient.count());
     }
 }
