@@ -1,15 +1,12 @@
 package example.user.config;
 
-import java.util.Locale;
-
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.web.servlet.LocaleResolver;
-import org.springframework.web.servlet.i18n.SessionLocaleResolver;
+import org.springframework.web.filter.CommonsRequestLoggingFilter;
 
 @Configuration
 @ComponentScan(basePackages = { "example.user" })
@@ -17,14 +14,12 @@ import org.springframework.web.servlet.i18n.SessionLocaleResolver;
 @EnableFeignClients(basePackages = { "example.user.service" })
 public class ApplicationConfig {
     @Bean
-    public LocaleResolver localeResolver() {
-        SessionLocaleResolver slr = new SessionLocaleResolver();
-        slr.setDefaultLocale(Locale.ENGLISH);
-        return slr;
+    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
+        return new PropertySourcesPlaceholderConfigurer();
     }
 
     @Bean
-    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
-        return new PropertySourcesPlaceholderConfigurer();
+    public CommonsRequestLoggingFilter requestLoggingFilter() {
+        return new RequestLoggingFilter();
     }
 }
