@@ -7,11 +7,14 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 
 @ControllerAdvice
 public class FooterController {
-    @Value("${runtime.environment}")
+    @Value("${runtime.environment:dev}")
     private String environment;
+    @Value("${info.app.version:-}")
+    private String version;
 
     @ModelAttribute
     public void setEnvironment(Model model) {
         model.addAttribute("environment", environment);
+        model.addAttribute("version", version);
     }
 }
